Version 11.2.12-pre

+++ Src/OSF/abseil/absl/debugging/internal/demangle.cc
@@ -1,22 +1,11 @@
 // Copyright 2018 The Abseil Authors.
-//
 // Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
 //
-//      https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 // For reference check out:
 // https://itanium-cxx-abi.github.io/cxx-abi/abi.html#mangling
 //
 // Note that we only have partial C++11 support yet.
-
+//
 #include "absl/absl-internal.h"
 #pragma hdrstop
 #include "absl/debugging/internal/demangle.h"
@@ -856,7 +845,7 @@ static bool ParseNumber(State * state, int * number_out) {
 		state->parse_state.mangled_idx += p - RemainingInput(state);
 		if(number_out != nullptr) {
 			// Note: possibly truncate "number".
-			*number_out = number;
+			*number_out = static_cast<int>(number);
 		}
 		return true;
 	}
