+++ Src/OSF/abseil/absl/profiling/internal/exponential_biased.cc
@@ -1,17 +1,6 @@
 // Copyright 2019 The Abseil Authors.
-//
 // Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
 //
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 #include "absl/absl-internal.h"
 #pragma hdrstop
 #include "absl/profiling/internal/exponential_biased.h"
@@ -29,14 +18,13 @@ namespace profiling_internal {
 // -log_e(q)/m = x
 // log_2(q) * (-log_e(2) * 1/m) = x
 // In the code, q is actually in the range 1 to 2**26, hence the -26 below
-int64_t ExponentialBiased::GetSkipCount(int64_t mean) {
+int64_t ExponentialBiased::GetSkipCount(int64_t mean) 
+{
 	if(ABSL_PREDICT_FALSE(!initialized_)) {
 		Initialize();
 	}
-
 	uint64_t rng = NextRandom(rng_);
 	rng_ = rng;
-
 	// Take the top 26 bits as the random number
 	// (This plus the 1<<58 sampling bound give a max possible step of
 	// 5194297183973780480 bytes.)
@@ -57,12 +45,10 @@ int64_t ExponentialBiased::GetSkipCount(int64_t mean) {
 	}
 	double value = std::rint(interval);
 	bias_ = interval - value;
-	return value;
+	return static_cast<int64_t>(value);
 }
 
-int64_t ExponentialBiased::GetStride(int64_t mean) {
-	return GetSkipCount(mean - 1) + 1;
-}
+int64_t ExponentialBiased::GetStride(int64_t mean) { return GetSkipCount(mean - 1) + 1; }
 
 void ExponentialBiased::Initialize() {
 	// We don't get well distributed numbers from `this` so we call NextRandom() a
