+++ Src/OSF/abseil/absl/strings/numbers.cc
@@ -1,20 +1,8 @@
 // Copyright 2017 The Abseil Authors.
-//
 // Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
 //
-//      https://www.apache.org/licenses/LICENSE-2.0
+// This file contains string processing functions related to numeric values.
 //
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-// This file contains string processing functions related to
-// numeric values.
-
 #include "absl/absl-internal.h"
 #pragma hdrstop
 #include "absl/strings/escaping.h"
@@ -386,7 +374,7 @@ static ExpDigits SplitToSix(const double value) {
 	// Since we'd like to know if the fractional part of d is close to a half,
 	// we multiply it by 65536 and see if the fractional part is close to 32768.
 	// (The number doesn't have to be a power of two,but powers of two are faster)
-	uint64_t d64k = d * 65536;
+	uint64_t d64k = static_cast<uint64_t>(d * 65536);
 	int dddddd; // A 6-digit decimal integer.
 	if((d64k % 65536) == 32767 || (d64k % 65536) == 32768) {
 		// OK, it's fairly likely that precision was lost above, which is
@@ -400,7 +388,7 @@ static ExpDigits SplitToSix(const double value) {
 		// value we're representing, of course, is M.mmm... * 2^exp2.
 		int exp2;
 		double m = std::frexp(value, &exp2);
-		uint64_t mantissa = m * (32768.0 * 65536.0 * 65536.0 * 65536.0);
+		uint64_t mantissa = static_cast<uint64_t>(m * (32768.0 * 65536.0 * 65536.0 * 65536.0));
 		// std::frexp returns an m value in the range [0.5, 1.0), however we
 		// can't multiply it by 2^64 and convert to an integer because some FPUs
 		// throw an exception when converting an number higher than 2^63 into an
